+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Fixed
 - 'Unreleased' link in this changelog
 - In runner tests wait result from processes for up to 0.1 seconds
+- Runner didn't wait for unix socket
 
 ## 0.1.0 - 2018-01-28
 ### Added